Lots of things! Read further :
Fixed a bug with dragging where the bounds were considered to be
Rectangle(top,left,bottom,right) whereas it should be
Rectangle(x,y,width,height)

Added the possibility to clone the corresponding HTMLElement when
cloning a DOMElement. A wrapperOnly property now controls whether or not
the HTMLElement should be rendered using the DOMElement's transformation
matrix.

The DOMElement's visibility now depends on whether or not it has a
parent that is visible on the stage.

Fixed a bug for shapes where the isVisible method would return the
graphics instance instead of a Boolean value.

Added a MOUSE_WHEEL event that works the same as the one in Flash.

Manually merged commit e1025c69df288e18797c7d628d1268c6151e055a and
2ec164399fd4b5e2528903bc0fa906fd9c391108 from the original repo.